optimizes eager loading in :zeitwerk mode
During initialization, the eager load paths of engines are unshifted
into AS::Dependencies.autoload_paths. After that, the collection is
frozen. (See the initializers in railties/lib/rails/engine.rb.)

Therefore, there is no eager load path that is not in the autoload
paths, and so the array difference in the deleted code is returning
always empty arrays. Just do nothing.